implem data parser

+++ source/fr/polytech/unice/exception/OverLoadedBinException.java
@@ -1,7 +1,7 @@
 package fr.polytech.unice.exception;
 
 public class OverLoadedBinException extends Exception {
-    public OverLoadedBinException(double freeSpace, double load) {
+    public OverLoadedBinException(int freeSpace, int load) {
         super("Espace libre : " + freeSpace + ", charge : " + load);
     }
 }
